Configure for multiple runs/restarts
Add $year to cable.nml (the payu CABLE driver will set this to the
appropriate year value corresponding to the restart file).

Note, currently we only have met forcing files up to year 1909 so the
current configuration will break if we run past this point. Should we
add the functionality to configure repeat forcing?